Remove support for disabling floating point numbers in generated C code
The way we want the x_is_in_range() functions to work requires the
floating point numbers code to

+++ tests/files/c_source/choices.c
@@ -95,7 +95,7 @@ int choices_foo_init(struct choices_foo_t *msg_p)
 
 int8_t choices_foo_foo_encode(double value)
 {
-    return (int8_t)(value);
+    return (int8_t)((int64_t)value);
 }
 
 double choices_foo_foo_decode(int8_t value)
