+++ tests/files/c_source/motohawk_sender_node.c
@@ -130,7 +130,7 @@ int motohawk_example_message_init(struct motohawk_example_message_t *msg_p)
 
 uint8_t motohawk_example_message_enable_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 bool motohawk_example_message_enable_is_in_range(uint8_t value)
@@ -140,7 +140,7 @@ bool motohawk_example_message_enable_is_in_range(uint8_t value)
 
 uint8_t motohawk_example_message_average_radius_encode(double value)
 {
-    return (uint8_t)(value / 0.1);
+    return (uint8_t)((int64_t)(value / 0.1));
 }
 
 bool motohawk_example_message_average_radius_is_in_range(uint8_t value)
@@ -150,7 +150,7 @@ bool motohawk_example_message_average_radius_is_in_range(uint8_t value)
 
 int16_t motohawk_example_message_temperature_encode(double value)
 {
-    return (int16_t)((value - 250.0) / 0.01);
+    return (int16_t)((int64_t)((value - 250.0) / 0.01));
 }
 
 double motohawk_example_message_temperature_decode(int16_t value)
