+++ tests/files/c_source/multiplex.c
@@ -157,7 +157,7 @@ int multiplex_message1_init(struct multiplex_message1_t *msg_p)
 
 uint8_t multiplex_message1_multiplexor_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double multiplex_message1_multiplexor_decode(uint8_t value)
@@ -172,7 +172,7 @@ bool multiplex_message1_multiplexor_is_in_range(uint8_t value)
 
 uint8_t multiplex_message1_bit_j_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double multiplex_message1_bit_j_decode(uint8_t value)
@@ -187,7 +187,7 @@ bool multiplex_message1_bit_j_is_in_range(uint8_t value)
 
 uint8_t multiplex_message1_bit_c_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double multiplex_message1_bit_c_decode(uint8_t value)
@@ -202,7 +202,7 @@ bool multiplex_message1_bit_c_is_in_range(uint8_t value)
 
 uint8_t multiplex_message1_bit_g_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double multiplex_message1_bit_g_decode(uint8_t value)
@@ -217,7 +217,7 @@ bool multiplex_message1_bit_g_is_in_range(uint8_t value)
 
 uint8_t multiplex_message1_bit_l_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double multiplex_message1_bit_l_decode(uint8_t value)
@@ -232,7 +232,7 @@ bool multiplex_message1_bit_l_is_in_range(uint8_t value)
 
 uint8_t multiplex_message1_bit_a_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double multiplex_message1_bit_a_decode(uint8_t value)
@@ -247,7 +247,7 @@ bool multiplex_message1_bit_a_is_in_range(uint8_t value)
 
 uint8_t multiplex_message1_bit_k_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double multiplex_message1_bit_k_decode(uint8_t value)
@@ -262,7 +262,7 @@ bool multiplex_message1_bit_k_is_in_range(uint8_t value)
 
 uint8_t multiplex_message1_bit_e_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double multiplex_message1_bit_e_decode(uint8_t value)
@@ -277,7 +277,7 @@ bool multiplex_message1_bit_e_is_in_range(uint8_t value)
 
 uint8_t multiplex_message1_bit_d_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double multiplex_message1_bit_d_decode(uint8_t value)
@@ -292,7 +292,7 @@ bool multiplex_message1_bit_d_is_in_range(uint8_t value)
 
 uint8_t multiplex_message1_bit_b_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double multiplex_message1_bit_b_decode(uint8_t value)
@@ -307,7 +307,7 @@ bool multiplex_message1_bit_b_is_in_range(uint8_t value)
 
 uint8_t multiplex_message1_bit_h_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double multiplex_message1_bit_h_decode(uint8_t value)
@@ -322,7 +322,7 @@ bool multiplex_message1_bit_h_is_in_range(uint8_t value)
 
 uint8_t multiplex_message1_bit_f_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double multiplex_message1_bit_f_decode(uint8_t value)
