+++ tests/files/c_source/open_actuator.c
@@ -152,7 +152,7 @@ int open_actuator_control_cmd_init(struct open_actuator_control_cmd_t *msg_p)
 
 uint8_t open_actuator_control_cmd_crc8_cmd1_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double open_actuator_control_cmd_crc8_cmd1_decode(uint8_t value)
@@ -169,7 +169,7 @@ bool open_actuator_control_cmd_crc8_cmd1_is_in_range(uint8_t value)
 
 uint8_t open_actuator_control_cmd_target_mode_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double open_actuator_control_cmd_target_mode_decode(uint8_t value)
@@ -184,7 +184,7 @@ bool open_actuator_control_cmd_target_mode_is_in_range(uint8_t value)
 
 uint8_t open_actuator_control_cmd_target_motor_id_cmd1_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double open_actuator_control_cmd_target_motor_id_cmd1_decode(uint8_t value)
@@ -199,7 +199,7 @@ bool open_actuator_control_cmd_target_motor_id_cmd1_is_in_range(uint8_t value)
 
 int16_t open_actuator_control_cmd_position_cmd_64_encode(double value)
 {
-    return (int16_t)(value / 0.0154286);
+    return (int16_t)((int64_t)(value / 0.0154286));
 }
 
 double open_actuator_control_cmd_position_cmd_64_decode(int16_t value)
@@ -214,7 +214,7 @@ bool open_actuator_control_cmd_position_cmd_64_is_in_range(int16_t value)
 
 int16_t open_actuator_control_cmd_torque_command_8_encode(double value)
 {
-    return (int16_t)(value / 0.0166667);
+    return (int16_t)((int64_t)(value / 0.0166667));
 }
 
 double open_actuator_control_cmd_torque_command_8_decode(int16_t value)
@@ -229,7 +229,7 @@ bool open_actuator_control_cmd_torque_command_8_is_in_range(int16_t value)
 
 uint8_t open_actuator_control_cmd_torque_close_loop_max_32_encode(double value)
 {
-    return (uint8_t)(value / 0.186666);
+    return (uint8_t)((int64_t)(value / 0.186666));
 }
 
 double open_actuator_control_cmd_torque_close_loop_max_32_decode(uint8_t value)
@@ -244,7 +244,7 @@ bool open_actuator_control_cmd_torque_close_loop_max_32_is_in_range(uint8_t valu
 
 uint8_t open_actuator_control_cmd_counter_cmd1_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double open_actuator_control_cmd_counter_cmd1_decode(uint8_t value)
@@ -308,7 +308,7 @@ int open_actuator_limits_cmd_init(struct open_actuator_limits_cmd_t *msg_p)
 
 uint8_t open_actuator_limits_cmd_crc8_cmd2_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double open_actuator_limits_cmd_crc8_cmd2_decode(uint8_t value)
@@ -325,7 +325,7 @@ bool open_actuator_limits_cmd_crc8_cmd2_is_in_range(uint8_t value)
 
 uint8_t open_actuator_limits_cmd_counter_cmd2_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double open_actuator_limits_cmd_counter_cmd2_decode(uint8_t value)
@@ -340,7 +340,7 @@ bool open_actuator_limits_cmd_counter_cmd2_is_in_range(uint8_t value)
 
 uint16_t open_actuator_limits_cmd_velocity_limit_encode(double value)
 {
-    return (uint16_t)(value);
+    return (uint16_t)((int64_t)value);
 }
 
 double open_actuator_limits_cmd_velocity_limit_decode(uint16_t value)
@@ -357,7 +357,7 @@ bool open_actuator_limits_cmd_velocity_limit_is_in_range(uint16_t value)
 
 uint16_t open_actuator_limits_cmd_accel_limit_encode(double value)
 {
-    return (uint16_t)(value);
+    return (uint16_t)((int64_t)value);
 }
 
 double open_actuator_limits_cmd_accel_limit_decode(uint16_t value)
@@ -432,7 +432,7 @@ int open_actuator_control_status_init(struct open_actuator_control_status_t *msg
 
 uint8_t open_actuator_control_status_crc8_stat1_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double open_actuator_control_status_crc8_stat1_decode(uint8_t value)
@@ -449,7 +449,7 @@ bool open_actuator_control_status_crc8_stat1_is_in_range(uint8_t value)
 
 uint8_t open_actuator_control_status_counter_stat1_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double open_actuator_control_status_counter_stat1_decode(uint8_t value)
@@ -464,7 +464,7 @@ bool open_actuator_control_status_counter_stat1_is_in_range(uint8_t value)
 
 int16_t open_actuator_control_status_torque_actual_encode(double value)
 {
-    return (int16_t)(value / 0.015625);
+    return (int16_t)((int64_t)(value / 0.015625));
 }
 
 double open_actuator_control_status_torque_actual_decode(int16_t value)
@@ -479,7 +479,7 @@ bool open_actuator_control_status_torque_actual_is_in_range(int16_t value)
 
 uint8_t open_actuator_control_status_torque_close_loop_actual_encode(double value)
 {
-    return (uint8_t)(value / 0.125);
+    return (uint8_t)((int64_t)(value / 0.125));
 }
 
 double open_actuator_control_status_torque_close_loop_actual_decode(uint8_t value)
@@ -537,7 +537,7 @@ int open_actuator_system_status_init(struct open_actuator_system_status_t *msg_p
 
 uint8_t open_actuator_system_status_crc8_stat2_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double open_actuator_system_status_crc8_stat2_decode(uint8_t value)
@@ -554,7 +554,7 @@ bool open_actuator_system_status_crc8_stat2_is_in_range(uint8_t value)
 
 uint8_t open_actuator_system_status_counter_stat2_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double open_actuator_system_status_counter_stat2_decode(uint8_t value)
@@ -569,7 +569,7 @@ bool open_actuator_system_status_counter_stat2_is_in_range(uint8_t value)
 
 uint8_t open_actuator_system_status_chip_temp_encode(double value)
 {
-    return (uint8_t)(value - -60.0);
+    return (uint8_t)((int64_t)(value - -60.0));
 }
 
 double open_actuator_system_status_chip_temp_decode(uint8_t value)
@@ -642,7 +642,7 @@ int open_actuator_torque_sensor_data_init(struct open_actuator_torque_sensor_dat
 
 uint8_t open_actuator_torque_sensor_data_crc8_data1_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double open_actuator_torque_sensor_data_crc8_data1_decode(uint8_t value)
@@ -659,7 +659,7 @@ bool open_actuator_torque_sensor_data_crc8_data1_is_in_range(uint8_t value)
 
 uint8_t open_actuator_torque_sensor_data_counter_data1_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double open_actuator_torque_sensor_data_counter_data1_decode(uint8_t value)
@@ -674,7 +674,7 @@ bool open_actuator_torque_sensor_data_counter_data1_is_in_range(uint8_t value)
 
 int16_t open_actuator_torque_sensor_data_torque_sense_encode(double value)
 {
-    return (int16_t)(value / 0.01);
+    return (int16_t)((int64_t)(value / 0.01));
 }
 
 double open_actuator_torque_sensor_data_torque_sense_decode(int16_t value)
