+++ tests/files/c_source/open_actuator_sender_node.c
@@ -315,7 +315,7 @@ int open_actuator_control_status_init(struct open_actuator_control_status_t *msg
 
 uint8_t open_actuator_control_status_crc8_stat1_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double open_actuator_control_status_crc8_stat1_decode(uint8_t value)
@@ -332,7 +332,7 @@ bool open_actuator_control_status_crc8_stat1_is_in_range(uint8_t value)
 
 uint8_t open_actuator_control_status_counter_stat1_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 bool open_actuator_control_status_counter_stat1_is_in_range(uint8_t value)
@@ -342,7 +342,7 @@ bool open_actuator_control_status_counter_stat1_is_in_range(uint8_t value)
 
 int16_t open_actuator_control_status_torque_actual_encode(double value)
 {
-    return (int16_t)(value / 0.015625);
+    return (int16_t)((int64_t)(value / 0.015625));
 }
 
 bool open_actuator_control_status_torque_actual_is_in_range(int16_t value)
@@ -352,7 +352,7 @@ bool open_actuator_control_status_torque_actual_is_in_range(int16_t value)
 
 uint8_t open_actuator_control_status_torque_close_loop_actual_encode(double value)
 {
-    return (uint8_t)(value / 0.125);
+    return (uint8_t)((int64_t)(value / 0.125));
 }
 
 bool open_actuator_control_status_torque_close_loop_actual_is_in_range(uint8_t value)
@@ -389,7 +389,7 @@ int open_actuator_system_status_init(struct open_actuator_system_status_t *msg_p
 
 uint8_t open_actuator_system_status_crc8_stat2_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 bool open_actuator_system_status_crc8_stat2_is_in_range(uint8_t value)
@@ -401,7 +401,7 @@ bool open_actuator_system_status_crc8_stat2_is_in_range(uint8_t value)
 
 uint8_t open_actuator_system_status_counter_stat2_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 bool open_actuator_system_status_counter_stat2_is_in_range(uint8_t value)
@@ -411,7 +411,7 @@ bool open_actuator_system_status_counter_stat2_is_in_range(uint8_t value)
 
 uint8_t open_actuator_system_status_chip_temp_encode(double value)
 {
-    return (uint8_t)(value - -60.0);
+    return (uint8_t)((int64_t)(value - -60.0));
 }
 
 bool open_actuator_system_status_chip_temp_is_in_range(uint8_t value)
