+++ tests/files/c_source/padding_bit_order.c
@@ -163,7 +163,7 @@ int padding_bit_order_msg0_init(struct padding_bit_order_msg0_t *msg_p)
 
 uint8_t padding_bit_order_msg0_b_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double padding_bit_order_msg0_b_decode(uint8_t value)
@@ -178,7 +178,7 @@ bool padding_bit_order_msg0_b_is_in_range(uint8_t value)
 
 uint16_t padding_bit_order_msg0_a_encode(double value)
 {
-    return (uint16_t)(value);
+    return (uint16_t)((int64_t)value);
 }
 
 double padding_bit_order_msg0_a_decode(uint16_t value)
@@ -193,7 +193,7 @@ bool padding_bit_order_msg0_a_is_in_range(uint16_t value)
 
 uint8_t padding_bit_order_msg0_d_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double padding_bit_order_msg0_d_decode(uint8_t value)
@@ -208,7 +208,7 @@ bool padding_bit_order_msg0_d_is_in_range(uint8_t value)
 
 uint16_t padding_bit_order_msg0_c_encode(double value)
 {
-    return (uint16_t)(value);
+    return (uint16_t)((int64_t)value);
 }
 
 double padding_bit_order_msg0_c_decode(uint16_t value)
@@ -272,7 +272,7 @@ int padding_bit_order_msg1_init(struct padding_bit_order_msg1_t *msg_p)
 
 uint8_t padding_bit_order_msg1_e_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double padding_bit_order_msg1_e_decode(uint8_t value)
@@ -287,7 +287,7 @@ bool padding_bit_order_msg1_e_is_in_range(uint8_t value)
 
 uint16_t padding_bit_order_msg1_f_encode(double value)
 {
-    return (uint16_t)(value);
+    return (uint16_t)((int64_t)value);
 }
 
 double padding_bit_order_msg1_f_decode(uint16_t value)
@@ -302,7 +302,7 @@ bool padding_bit_order_msg1_f_is_in_range(uint16_t value)
 
 uint8_t padding_bit_order_msg1_g_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double padding_bit_order_msg1_g_decode(uint8_t value)
@@ -317,7 +317,7 @@ bool padding_bit_order_msg1_g_is_in_range(uint8_t value)
 
 uint16_t padding_bit_order_msg1_h_encode(double value)
 {
-    return (uint16_t)(value);
+    return (uint16_t)((int64_t)value);
 }
 
 double padding_bit_order_msg1_h_decode(uint16_t value)
@@ -375,7 +375,7 @@ int padding_bit_order_msg2_init(struct padding_bit_order_msg2_t *msg_p)
 
 uint8_t padding_bit_order_msg2_i_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double padding_bit_order_msg2_i_decode(uint8_t value)
@@ -390,7 +390,7 @@ bool padding_bit_order_msg2_i_is_in_range(uint8_t value)
 
 uint8_t padding_bit_order_msg2_j_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double padding_bit_order_msg2_j_decode(uint8_t value)
@@ -405,7 +405,7 @@ bool padding_bit_order_msg2_j_is_in_range(uint8_t value)
 
 uint8_t padding_bit_order_msg2_k_encode(double value)
 {
-    return (uint8_t)(value);
+    return (uint8_t)((int64_t)value);
 }
 
 double padding_bit_order_msg2_k_decode(uint8_t value)
@@ -473,7 +473,7 @@ int padding_bit_order_msg3_init(struct padding_bit_order_msg3_t *msg_p)
 
 uint64_t padding_bit_order_msg3_l_encode(double value)
 {
-    return (uint64_t)(value);
+    return (uint64_t)((int64_t)value);
 }
 
 double padding_bit_order_msg3_l_decode(uint64_t value)
@@ -543,7 +543,7 @@ int padding_bit_order_msg4_init(struct padding_bit_order_msg4_t *msg_p)
 
 uint64_t padding_bit_order_msg4_m_encode(double value)
 {
-    return (uint64_t)(value);
+    return (uint64_t)((int64_t)value);
 }
 
 double padding_bit_order_msg4_m_decode(uint64_t value)
